pyproject.toml
[project]
name = "backend-practice"
version = "0.1.0"
description = ""
authors = [
    {name = "donggyu"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "dependency-injector (>=4.46.0,<5.0.0)",
    "bcrypt (==4.0.1)",
    "python-jose (>=3.4.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "passlib (>=1.7.4,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.2,<4.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=1.0.0,<2.0.0)",
    "pytest-mock (>=3.14.1,<4.0.0)",
    "python-ulid (>=3.0.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
